farm/requirements.txt
Adafruit-Blinka==8.18.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-lis3dh==5.1.22
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.1
adafruit-io==2.7.0
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
arandr==0.1.10
asyncio==3.4.3
automationhat==0.2.0
autopep8==2.0.2
azure-iot-device==2.12.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
dbus-python==1.2.16
deprecation==2.1.0
distro==1.5.0
dotenv==0.0.5
drumhat==0.1.0
envirophat==1.0.0
evdev==1.4.0
ExplorerHAT==0.4.2
Flask==1.1.2
fourletterphat==0.1.0
geographiclib==2.0
geopy==2.3.0
gpiozero==1.6.2
grove.py @ file:///home/braeden.local/Downloads/grove.py
html5lib==1.1
idna==3.4
itsdangerous==1.1.0
janus==1.0.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
MouseInfo==0.1.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
packaging==23.0
paho-mqtt==1.6.1
pantilthat==0.0.7
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.10.0
pyftdi==0.54.0
pygame==1.9.6
PyGetWindow==0.0.9
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
PyMsgBox==1.0.9
pynmea2==1.19.0
pyOpenSSL==20.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5b0
PySocks==1.7.1
python-apt==2.2.1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.4
pyusb==1.2.1
rainbowhat==0.1.0
requests==2.28.2
requests-oauthlib==1.0.0
requests-unixsocket==0.3.0
responses==0.12.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
seeed-python-reterminal==0.5
sense-hat==2.4.0
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
smbus2==0.4.2
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
tomli==2.0.1
touchphat==0.0.1
twython==3.8.2
typing-extensions==4.5.0
unicornhathd==0.0.4
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==1.0.1
